feat(factorization): реализована мультистратегическая факторизация - Добавлен новый класс src/Tools/Factorizer.php для факторизации. - Класс Factorizer поддерживает FFI, команду factor и удаленный сервис. - Обновлен C++ модуль primemodule.cpp с улучшенным алгоритмом Полларда-Брента. - Изменен FFI-интерфейс factorizeFFI на int64_t для лучшей совместимости. - Добавлена обработка ошибок и возвращение 0 или -1 в FFI-функции. - Исправлена ошибка парсинга pq TL-строки в MTProto/Client::parseResPQ. - Улучшена надежность и читаемость парсинга resPQ ответов.

+++ src/cpp/primemodule.h
@@ -1,9 +1,14 @@
+// File: src/cpp/primemodule.h
+// Он же используется как primemodule-ffi.h
+
 #ifndef FACTORIZATION_H
 #define FACTORIZATION_H
 
 #include <cstdint>
 
-// Прототип функции для факторизации числа
-extern "C" int32_t factorizeFFI(const char* number);
+// Прототип функции для факторизации числа.
+// Возвращает int64_t для лучшей совместимости с FFI.
+// 0 - неудача, -1 - ошибка парсинга, > 1 - найденный множитель.
+extern "C" int64_t factorizeFFI(const char* number);
 
 #endif // FACTORIZATION_H
